+++ src/openepd/api/sync_client.py
@@ -17,6 +17,8 @@ __all__ = ("OpenEpdApiClientSync",)
 
 from requests.auth import AuthBase
 
+from openepd.api.average_dataset.generic_estimate_sync_api import GenericEstimateApi
+from openepd.api.average_dataset.industry_epd_sync_api import IndustryEpdApi
 from openepd.api.base_sync_client import SyncHttpClient, TokenAuth
 from openepd.api.category.sync_api import CategoryApi
 from openepd.api.epd.sync_api import EpdApi
@@ -40,6 +42,8 @@ class OpenEpdApiClientSync:
         self.__epd_api: EpdApi | None = None
         self.__pcr_api: PcrApi | None = None
         self.__category_api: CategoryApi | None = None
+        self.__generic_estimate_api: GenericEstimateApi | None = None
+        self.__industry_epd_api: IndustryEpdApi | None = None
 
     @property
     def epds(self) -> EpdApi:
@@ -61,3 +65,17 @@ class OpenEpdApiClientSync:
         if self.__category_api is None:
             self.__category_api = CategoryApi(self._http_client)
         return self.__category_api
+
+    @property
+    def industry_epds(self) -> IndustryEpdApi:
+        """Get the Category API."""
+        if self.__industry_epd_api is None:
+            self.__industry_epd_api = IndustryEpdApi(self._http_client)
+        return self.__industry_epd_api
+
+    @property
+    def generic_estimates(self) -> GenericEstimateApi:
+        """Get the GE API."""
+        if self.__generic_estimate_api is None:
+            self.__generic_estimate_api = GenericEstimateApi(self._http_client)
+        return self.__generic_estimate_api